ref(profiling): extract shouldProfileTransaction and start profiler in SDK init
